src/requirements.txt
pandas
numpy
matplotlib
wfdb